+++ test/public_api/scene_builder_test.dart
@@ -61,7 +61,7 @@ void _expectSameSceneDataContract(
 
 SceneSnapshot _duplicateNodeSnapshotFromInternalBypass() {
   return unsafeMaterializeSceneSnapshot(
-    sceneSnapshotBackingFromValidated(
+    SceneSnapshotBacking(
       backgroundLayer: backgroundLayerSnapshotBackingFromValidated(
         nodes: <NodeSnapshotBacking>[
           nodeSnapshotBackingOf(